Remove the extra 120 seconds of file time fudge.
This caused problems when an interval starts >0 and <120 seconds from a dump file.
The downside is that now when the file **prior** to the interval start is dumped early (e.g., the 23:55 file is dumped at 23:54:59) AND it ALSO contains more data than expected (e.g., it contains data for the 00:00:00 second), then that data will be missed.
The correct way to fix this issue is to move the duration field into the data table, and then set initialTime and duration based on the actual data in the file rather than on the nominal dump time and duration.